requirements.txt
grpcio==1.33.2
protobuf==3.14.0
PyBluez==0.23
RPi.GPIO==0.7.0
six==1.15.0
